+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Mediation/AppLovin/Platforms/iOS/Externs.cs
@@ -17,7 +17,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace GoogleMobileAds.iOS.Mediation.AppLovin
+namespace GoogleMobileAds.Mediation.AppLovin.iOS
 {
     // Externs used by the iOS component.
     internal class Externs